Implement a first draft of the websocket server
Implement a (draft for a) simple websocket-based protocol which can be
used by a frontend to interact with quinte server. Since websockets have
no "request context" or paths or query parameters but are merely a data
stream cut into frames, we have to provide some metadata in the "frames"
themselves. The protocol is JSON based (websockets + json make for an
easy interaction with browser-based frontend) and uses websocket's text
frames.

Every frame contains one JSON object, that always has two fields
* cid: the correlation ID is chosen by the party where the frame
  originates and includes in all direct responses to that frame,
  unaltered, uninterpreted.
* payload: contains whatever payload should be transmitted.

There are payload types MailSearch and MailList that can be used to
perform a search.

Everything is unstable. Everything is just to proof the concept.